refactor: move sidebar component to shared
